requirements.txt
cffi==1.17.1
colorama==0.4.6
packaging==24.2
pillow==11.1.0
pillow_heif==0.21.0
pip-review==1.3.0
pycparser==2.22
tqdm==4.67.1
